+++ src/lib/gen/gen.ts
@@ -3,7 +3,7 @@ import {join, basename, dirname} from 'node:path';
 import {z} from 'zod';
 
 import type {Filesystem} from '../fs/filesystem.js';
-import {genModuleMeta, toGenModuleType} from './genModule.js';
+import {gen_module_meta, to_gen_module_type} from './gen_module.js';
 import type {SourceId} from '../path/paths.js';
 
 export type GenResult = {
@@ -104,7 +104,7 @@ const to_output_file_id = (origin_id: SourceId, raw_file_name: string | undefine
 };
 
 export const toOutputFileName = (filename: string): string => {
-	const {pattern, text} = genModuleMeta[toGenModuleType(filename)];
+	const {pattern, text} = gen_module_meta[to_gen_module_type(filename)];
 	const parts = filename.split('.');
 	const gen_pattern_index = parts.indexOf(text);
 	if (gen_pattern_index === -1) {