+++ CHANGELOG.md
@@ -94,7 +94,7 @@
 
 ## 0.75.1
 
-- export `toGenContextImports` from `lib/runGen`
+- export `to_gen_context_imports` from `lib/run_gen`
   ([commit](https://github.com/feltjs/gro/commit/980f75bf43f77cb3107b446c079cc401c88b3f94))
 
 ## 0.75.0